+++ package.json
@@ -30,7 +30,6 @@
     "lodash": "^4.17.21",
     "match-sorter": "^6.3.1",
     "material-react-table": "^2.0.4",
-    "node-fetch": "2.6.1",
     "react": "^18.2.0",
     "react-day-picker": "^8.9.1",
     "react-dom": "^18.2.0",
